Fix z.ss01 (z.alt), see #15
The issue here is a bit interesting, so I'm going to outline it here.

The original issue reads:

"The 'z' in Alternates is missing the main shape. It's just a dash at the moment."

It turns out that this is only true for the ttf files of the MontserratAlternates files.

MontserratAlternates is using the custom parameter "Rename Glyphs" to, amongs others,
rename 'z.ss01' to 'z'. But, 'z' is referenced as a component in `z.ss01`. Thus, after
renaming the reference is circular and consquently not exported.

Glyphs exports the otf versions correctly though, since there all components are outlined.
In this case, the z was a glyph with mixed components and paths, so glyphs could have
outlined the ttf before doing the renaming, but it didn't. Hence the fix is converting
the z component to an outline in the source file.