Regression tested and fonts rebuild w/ static instances
There are *a lot* of changed to this new version versus the previous version of Montserrat. It is hard to determine if they are desired. However, as it seems the work was previously approved (it would have to have been for the font to be hand hinted), I am optimistic that they're OK.